More reliable property capture mechanism
Capture changed properties in SnoopUI now instead of in PropertyGrid.
Occurs on CurrentSelection setter, before the selection changes to the
new value.  Now we should always have the changed properties hooked to
the right owning object.  Got rid of hack where PropertyGrid2 needed to
be able to find and call SnoopUI.