Fixing the white foreground issue on Methods and PowerShell tabs.
There are apps that change the styling of the default TextBlock ... to easily implement white text on a black background. Since Snoop injects into the Snooped application, it takes on this default styling ... and impacts controls in the main Snoop UI. We need to set manually set the Style of all TextBlock(s) to x:Null in order to avoid this problem.