Checking in the source code for the bootstrapper.
The managed injector bits require the Microsoft Visual C++ 2010 redistributables to be installed on your machine. This is the installer source code for the bootstrapper which downloads and installs the redists (x86 and x64) for MSVC++. Once the bootstrapper runs, it then launches Snoop.msi. So, if you want to skip the bootstrapper, just install Snoop.msi instead of Setup.exe.