Fix delve into binding expression
When you delve into a binding expresion an exception was being thrown
because we're setting the Value of a DP (the Target of the
PropertyInspector and possibly the Value of the PropertyInformation) to
a BindingExpression. The framework thinks you're trying to apply a
binding and it generates an exception. The fix is to wrap the
expressions.