use InvariantCulture to display values
Unfortunately there is no generic ToString overload that accepts a CultureInfo. To circumvent this problem, I have to highjack the current threads CurrentCulture for a short period in time, generate the string value and switch the culture back.
There may be other places where it would be better to use the InvariantCulture, but this is the one I was primarily concerned about.

Affected Property types are at least:
- DateTime values
- any numbers with a decimal separator (float, double, decimal)
- anything that uses those numbers as part of their toString() implementation, like Geometry.