Improve the dark theme.
Previously, there were issues with the contrast levels in the dark theme.
Here's an example using the alternate row background color "Snoop.Brushes.ItemsControl.AlternationBackground".
Contrast values come from the tool https://webaim.org/resources/contrastchecker/

Light theme:
Background: FFFFFF
Alternate: EAEAEA
Contrast: 1.2

Dark theme:
Background: 000000
Alternate: 757575
Contrast 4.55

Dark theme (with this change)
Background: 000000
Foreground: 151515
Contrast: 1.14

Without this change, the contrast between the background colors of two adjacent rows is 4.55, which is very distracting. That's the kind of contrast you would want between foreground and background, not between two adjacent parts of the background. With the change, the contrast is 1.14, which is very close to the contrast in the light theme of 1.2.