Fix NuGet packages not included during build of CompileOtherArches
When building InjectorLauncher for other architectures, I noticed that the NuGet packages such as ReferenceAssemblies.net452, SourceLink, etc. are not included.

This is because CompileOtherArches issues a single build of both Restore and Build targets. At Restore time, the generated obj\x64\Snoop.InjectorLauncher.csproj.nuget.g.targets doesn't exist yet (duh!), so it's not imported during evaluation for Restore.

At Build time, MSBuild reuses the previous evaluation used to build the Restore target, and that evaluation didn't import the generated nuget.g.targets file because it didn't exist yet.

The fix is to force the Build target to evaluate from scratch, so that it can pick up the target from disk again. We achieve this by declaring a dummy property that is not used anywhere. MSBuild will not reuse the previous evaluation as it doesn't have the exact same set of properties.

Now restoring and building on a clean machine works fine even if the reference assemblies targeting pack for .NET 4.5.2 isn't installed.

Not sure why RunEachTargetSeparately="false" was there as false is the default. Setting it to true didn't fix the issue as the evaluation was still reused.